refactor(core): one duration parser for every command surface
