+++ crates/homeboy-core/src/lib.rs
@@ -88,6 +88,7 @@ pub mod controller_runtime;
 pub use homeboy_lifecycle_contract::cook_status;
 pub mod daemon;
 pub mod deps;
+pub mod duration;
 pub mod engine;
 pub use homeboy_lab_contract::env_materialization_plan;
 // error moved to the internal `homeboy-error` crate. Re-exported here so existing
